refactor(ui): moved popups into one file
