+++ src/ui/atoms/index.js
@@ -7,4 +7,3 @@ export { Input } from "./input"
 export { Link } from "./link"
 export { Text } from "./text"
 export { Box } from "./box"
-export { PopUp } from "./popup"
